+++ src/package.rs
@@ -135,10 +135,10 @@ impl Package {
         Ok(())
     }
 
-    pub(crate) fn set_cargo_version(&self, v: Version) -> anyhow::Result<()> {
+    pub(crate) fn set_cargo_version(&self, v: Version, update_lock: bool) -> anyhow::Result<()> {
         let vs = v.to_string();
         self.set_package_field("version", &vs)?;
-        if self.is_bin() && self.path().join("Cargo.lock").exists() {
+        if update_lock {
             LoggedCommand::new("cargo")
                 .arg("update")
                 .arg("-p")
@@ -266,7 +266,7 @@ mod tests {
             "[dependencies]\n",
         ));
         tpkg.package
-            .set_cargo_version(Version::new(1, 2, 3))
+            .set_cargo_version(Version::new(1, 2, 3), false)
             .unwrap();
         tpkg.manifest.assert(concat!(
             "[package]\n",
@@ -282,7 +282,7 @@ mod tests {
     fn set_cargo_version_inline() {
         let tpkg = TestPackage::new("package = { name = \"foobar\", version = \"0.1.0\", edition = \"2021\" }\ndependencies = {}\n");
         tpkg.package
-            .set_cargo_version(Version::new(1, 2, 3))
+            .set_cargo_version(Version::new(1, 2, 3), false)
             .unwrap();
         tpkg.manifest.assert("package = { name = \"foobar\", version = \"1.2.3\", edition = \"2021\" }\ndependencies = {}\n");
     }
@@ -297,7 +297,7 @@ mod tests {
             "[dependencies]\n",
         ));
         tpkg.package
-            .set_cargo_version(Version::new(1, 2, 3))
+            .set_cargo_version(Version::new(1, 2, 3), false)
             .unwrap();
         tpkg.manifest.assert(concat!(
             "[package]\n",
@@ -315,7 +315,7 @@ mod tests {
         let tpkg = TestPackage::new("[dependencies]\n");
         assert!(tpkg
             .package
-            .set_cargo_version(Version::new(1, 2, 3))
+            .set_cargo_version(Version::new(1, 2, 3), false)
             .is_err());
         tpkg.manifest.assert("[dependencies]\n");
     }
@@ -326,7 +326,7 @@ mod tests {
         let tpkg = TestPackage::new("package = 42\n");
         assert!(tpkg
             .package
-            .set_cargo_version(Version::new(1, 2, 3))
+            .set_cargo_version(Version::new(1, 2, 3), false)
             .is_err());
         tpkg.manifest.assert("package = 42\n");
     }